requirements.txt
PyYAML
beautifulsoup4==4.9.0
lxml==4.5.0
